feat!: default capo sorting
