+++ docs/content/1.usage/2.guides/2.sorting.md
@@ -3,13 +3,15 @@ title: Tag Sorting
 description: How tags are sorted and how to configure them.
 ---
 
-Once tags are [deduped](/usage/guides/handling-duplicates), they will be sorted. Sorting the tags is important
-to ensure critical tags are rendered first, as well as allowing you to have tags in a specific order that you need them in.
+## Introduction
 
-For example, if you need to preload an asset, you'll need this to come before the asset itself. Which is a bit of a challenge
-when the tags are nested.
+Sorting the tags is important to ensure critical tags are rendered first, as well as allowing you to have tags in a specific order that you need them in.
 
-## Sorting Logic
+## Tag Sorting Logic
+
+Sorting is first done using the [Capo.js](https://rviscomi.github.io/capo.js/) weights, making sure tags are rendered in
+a specific way to avoid [Critical Request Chains](https://web.dev/critical-request-chains/) issues as well
+as rendering bugs.
 
 Sorting is done in multiple steps:
 - order critical tags first
