+++ packages/unhead/src/index.ts
@@ -18,6 +18,5 @@ export * from './composables/useSeoMeta'
 export * from './composables/useServerHead'
 export * from './composables/useServerHeadSafe'
 export * from './composables/useServerSeoMeta'
-export * from './optionalPlugins/capoPlugin'
 // optional plugins
 export * from './optionalPlugins/hashHydrationPlugin'
