+++ packages/vue/export-size-report.json
@@ -113,13 +113,6 @@
       "bundled": 1618
     },
     {
-      "name": "CapoPlugin",
-      "path": "dist/index.mjs",
-      "minified": 179,
-      "minzipped": 132,
-      "bundled": 460
-    },
-    {
       "name": "HashHydrationPlugin",
       "path": "dist/index.mjs",
       "minified": 188,
