+++ test/vue/ssr/optionsApi.test.ts
@@ -24,15 +24,13 @@ describe('vue ssr options api', () => {
       ],
     })
 
-    expect(headResult.headTags).toMatchInlineSnapshot(
-      `
+    expect(headResult.headTags).toMatchInlineSnapshot(`
       "<title>hello</title>
+      <script src="foo.js"></script>
       <meta name="description" content="desc">
       <meta property="og:locale:alternate" content="fr">
-      <meta property="og:locale:alternate" content="zh">
-      <script src="foo.js"></script>"
-    `,
-    )
+      <meta property="og:locale:alternate" content="zh">"
+    `)
     expect(headResult.htmlAttrs).toEqual(' lang="zh"')
   })
 })
